Move method to core.emails
